Copies over the old config (1bash, 1oc)
It copies over the old configured variables from 1bash and 1oc after an update
It reads 1bash_latest and 1oc_latest lines by lines and checked if it is a variable, if it is;
it reads 1bash and 1oc and find the same variable there.  If it isn't the same value, it copies it over.

To make it works 100% time, all the "DO NOT TOUCH" Variables must be cleaned up in another bash and only let the "configurable by user" Variables in the 1bash.